+++ public/js/controllers/voice.js
@@ -46,12 +46,14 @@ eva.controller('voice', ['$scope', '$http', function ($scope, $http) {
     }
 
     $scope.delete = function (id) {
-        $http.delete('/api/common/' + id + '?db=voice').then(function successCallback(response) {
-            $scope.list();
-            notify(locale().VOICE.NOTIFY.DELETE.SUCCESS);
-        }, function errorCallback(response) {
-            notify(locale().VOICE.NOTIFY.ERROR,  'danger');
-        });;
+        if (confirm(locale().COMMON.DELETE)) {
+            $http.delete('/api/common/' + id + '?db=voice').then(function successCallback(response) {
+                $scope.list();
+                notify(locale().VOICE.NOTIFY.DELETE.SUCCESS);
+            }, function errorCallback(response) {
+                notify(locale().VOICE.NOTIFY.ERROR, 'danger');
+            });
+        }
     }
 
     $scope.dataTable = function (way = 0) {
